Revert "Events lookup"
